pyproject.toml
[project]
name = "pymmcore-proxy"
version = "0.1.0"
description = "Network proxy for pymmcore-plus — control microscopes remotely"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "httpx",
    "starlette",
    "uvicorn",
    "websockets",
    "psygnal>=0.9",
    "useq-schema",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pymmcore-plus",
]
server = [
    "pymmcore-plus",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pymmcore_proxy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
